+++ src/components/merchantPage/MenuItemCard.jsx
@@ -2,6 +2,7 @@ import PropTypes from "prop-types";
 import Cookies from "js-cookie";
 
 const MenuItemCard = ({ food, onClick }) => {
+    // console.debug("MenuItemCard food:", food);
     const { name, picture, price, description } = food;
 
     const authToken = Cookies.get("authToken");
@@ -37,7 +38,7 @@ const MenuItemCard = ({ food, onClick }) => {
                     <p className="text-sm text-gray-600 mt-2 line-clamp-3 text-ellipsis">{description}</p>
 
                     {/* Add button */}
-                    {authToken && ( // 僅當 authToken 存在時顯示按鈕
+                    {authToken && ( // if user is logged in
                         <div className="flex justify-end mt-4 absolute bottom-[15px] right-[15px]">
                             <button className="bg-orange-500 text-white rounded-full w-10 h-10 flex items-center justify-center">
                                 <span className="text-2xl font-bold mb-1">+</span>
