+++ src/pages/Menu.jsx
@@ -3,6 +3,8 @@ import { useRef, useState, useEffect } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSpinner } from "@fortawesome/free-solid-svg-icons";
 import { useQuery, useQueries } from "@tanstack/react-query";
+import { useCategoryQueries } from "../hooks/menu/useCategoryQueries";
+import { useCategoryListQuery } from "../hooks/menu/useCategoryListQuery";
 import MenuHeader from "../components/merchantPage/MenuHeader";
 import MenuNavbar from "../components/merchantPage/MenuNavbar";
 import MenuSection from "../components/merchantPage/MenuSection";
@@ -18,6 +20,7 @@ function Menu() {
     const [isNavbarFixed, setIsNavbarFixed] = useState(false);
     const setNavbarItems = useNavStore((state) => state.setNavbarItems);
     const setDishes = useAllDishesStore((state) => state.setDishes);
+
     // handle scroll to section
     const handleScrollToSection = (index) => {
         sectionRefs.current[index]?.scrollIntoView({
@@ -64,64 +67,24 @@ function Menu() {
         }
     }, [merchantId, getMerchantById]);
 
-    // Fetch menu category list and dish details
-    const { data: menuCategoryList = [] } = useQuery({
-        queryKey: ["menuCategoryList", menuId],
-        queryFn: async () => {
-            const res = await getMenuClient.getMenuByMenuId(menuId);
-            const categories = res.data.categories;
-            // Update navbar items
-            setNavbarItems(categories.map((category) => category.first));
-            return categories;
-        },
-        enabled: menuId != undefined,
-        refetchOnWindowFocus: false,
-    });
+    const menuCategoryList = useCategoryListQuery(menuId);
+    setNavbarItems(menuCategoryList.map((category) => category.first));
+    const categoryData = useCategoryQueries(menuCategoryList, merchantId);
 
-    // Fetch dish details for each category separately
-    const categoryQueries = useQueries({
-        queries: menuCategoryList.map((category) => ({
-            queryKey: ["categoryDishes", merchantId, category.first],
-            queryFn: async () => {
-                const dishDetails = await getMenuClient.getDishsByCategory(merchantId, category.first);
-                return {
-                    categoryName: category.first,
-                    dishes: dishDetails,
-                };
-            },
-            enabled: menuId != undefined && !!category.second.length,
-            refetchOnWindowFocus: false,
-        })),
-    });
-    useEffect(() => {
-        categoryQueries.forEach((query) => {
-            if (query.isSuccess && query.data) {
-                // save dishes' data to allDishStore
-                const dishesToStore = query.data.dishes.reduce((acc, dish) => {
-                    acc[dish.id] = dish;
-                    return acc;
-                }, {});
-                setDishes(dishesToStore);
-            }
-        });
-    }, [categoryQueries, setDishes]);
+    // useEffect(() => {
+    //     categoryQueries.forEach((query) => {
+    //         if (query.isSuccess && query.data) {
+    //             // save dishes' data to allDishStore
+    //             const dishesToStore = query.data.dishes.reduce((acc, dish) => {
+    //                 acc[dish.id] = dish;
+    //                 return acc;
+    //             }, {});
+    //             setDishes(dishesToStore);
+    //         }
+    //     });
+    // }, [categoryQueries, setDishes]);
 
-    useEffect(() => {
-        categoryQueries.forEach((query) => {
-            if (query.isSuccess && query.data) {
-                // save dishes' data to allDishStore
-                const dishesToStore = query.data.dishes.reduce((acc, dish) => {
-                    acc[dish.id] = dish;
-                    return acc;
-                }, {});
-                setDishes(dishesToStore);
-            }
-        });
-    }, [categoryQueries, setDishes]);
-    // Transform the queries results into categoryData
-    const categoryData = categoryQueries
-        .map((query) => query.data)
-        .filter(Boolean); // Filter out undefined results
+
     return merchant && merchantId ? (
         <div>
             <MenuHeader merchantData={merchant} />
